src/rag/requirements.txt
langchain-core>=0.1.27
langchain-community>=0.0.24
langchain-chroma>=0.0.10
langchain-huggingface>=0.0.10
openai>=1.12.0
tenacity>=8.2.3
chromadb>=0.4.24
sentence-transformers>=2.5.1
pyyaml>=6.0.1
python-dotenv>=1.0.1 